LayoutTests/fast/forms/month/month-stepup-stepdown-from-renderer-expected.txt

Check stepping-up and -down for month input fields from renderer. No cases of empty initial values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Function arguments are (value, step, {min or max}, [stepCount]).
Normal cases
PASS stepUp("2010-02", null, null) is "2010-03"
PASS stepDown("2010-02", null, null) is "2010-01"
PASS stepUp("2010-02", null, null, 10) is "2010-12"
PASS stepDown("2010-02", null, null, 11) is "2009-03"
PASS stepUp("1970-01", "4", null, 2) is "1970-09"
PASS stepDown("1970-01", "4", null, 3) is "1969-01"
Step=any
PASS stepUp("2010-02", "any", null) is "2010-03"
PASS stepDown("2010-02", "any", null) is "2010-01"
Overflow/underflow
PASS stepUp("2010-02", "3.40282346e+38", null) is "275760-09"
PASS stepDown("2010-02", "3.40282346e+38", null) is "1970-01"
PASS stepUp("2010-02", "1", "2010-02") is "2010-02"
PASS stepDown("2010-02", "1", "2010-02") is "2010-02"
stepDown()/stepUp() for stepMismatch values
PASS stepDown("2010-02", "3", "2009-10") is "2010-01"
PASS stepUp("1970-02", "4", "") is "1970-05"

PASS successfullyParsed is true

TEST COMPLETE
